auth token added to menuitemview

+++ littlelemon/restaurant/models.py
@@ -9,4 +9,4 @@ class Booking(models.Model):
 class Menu(models.Model):
     title=models.CharField(max_length=255)
     price=models.DecimalField(max_digits=10,decimal_places=2)
-    inventory=models.IntegerField()
+    inventory=models.SmallIntegerField()
